# Skip leadership check if the etcd instance is active processing heartbeats