Resolve https://github.com/etcd-io/etcd/issues/18069

Will post benchmark result later.


Please read https://github.com/etcd-io/etcd/blob/main/CONTRIBUTING.md#contribution-flow.