@ahrtr 2024-08-09
> I have one question: Imagine a scenario where leadership is lost due to two candidates fighting over with similar terms and equal votes.
> Does that mean after merging this the new vote will be start with more delay? And possibly the cluster will be down?

This PR has nothing to do with any raft's protocol. It just tries to avoid unnecessary `ensureLeadership` check, so as to avoid unnecessary performance reduce.