@ahrtr 2024-08-15
> The proposed solution is not compatible with etcd configuration options and doesn't work for 5 node clusters. Please consider reverting and redesign.

Don't agree. The only minor possible improvement is to guard the threshold < electionTimeout, but I am not worry about it as mentioned in https://github.com/etcd-io/etcd/pull/18428#discussion_r1718609232